Modbus_Proxy/test/integration/requirements.txt
requests>=2.28.0
paho-mqtt>=1.6.0
pytest>=7.0.0
pytest-timeout>=2.1.0
